refactor(assets): Rebuild asset structure based on features
