+++ src/components/common/icons/SearchIcon.tsx
@@ -1,4 +1,4 @@
-import Search from "../../../assets/figmaIcon/32px/Search.svg?react";
+import Search from "../../../assets/search-input/search.svg?react";
 
 interface SearchIconProps {
   className?: string;
